pyproject.toml
[project]
name = "agent-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "langchain>=1.2.9",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.8",
    "langgraph>=1.0.8",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.1",
    "redis-mcp-server>=0.4.1",
    "sse-starlette>=3.2.0",
    "starlette>=0.52.1",
    "uvicorn>=0.40.0",
]
